requirements.txt
# Phase 1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings
email-validator>=2.0.0
typer==0.9.0
click>=8.1.7
rich>=13.7.0
sqlalchemy>=1.4,<2.0
alembic==1.13.1
psycopg2-binary>=2.9.9
redis==5.0.1
celery==5.3.4
#apache-airflow==2.7.2
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
torch>=2.1.1
transformers>=4.35.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
openai==1.3.8
httpx==0.25.2
aiofiles==23.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv>=1.0.0
pyyaml==6.0.1
structlog==23.2.0
types-requests>=2.31.0

# Phase 2 Ingestion connectors
psycopg[binary]>=3.1.0
aiomysql==0.2.0
aiosqlite==0.19.0
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0
python-magic==0.4.27
aiohttp==3.9.1
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.1
google-cloud-bigquery>=3.15.0
google-auth>=2.25.0
redshift-connector>=2.0.915
databricks-sql-connector>=2.9.3
clickhouse-driver>=0.2.6
cx-Oracle>=8.3.0
pyodbc>=5.0.1
types-aiofiles>=23.2.0

# Phase 3 NLP and Processing
spacy==3.7.4
langdetect==1.0.9
beautifulsoup4==4.12.3
nltk>=3.9.1
pydantic-settings>=2.5.0

# Phase 4  Embeddings and Vector Stores
faiss-cpu>=1.8.0
qdrant-client==1.9.0
scikit-learn>=1.3.2

# Phase 5 additions - API Serving and LLM Integration
fastapi>=0.111.0
pydantic==2.7.1
Jinja2==3.1.4
accelerate>=0.30.1
bitsandbytes>=0.42.0
cohere>=4.38
anthropic==0.25.4
google-generativeai==0.3.2
pickle-mixin>=1.0.2
starlette>=0.37.2
asyncio-throttle>=1.0.2
types-redis>=4.6
